requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-command==0.9.1
pulumi-docker==4.4.4
pulumi-tailscale==0.13.2
pulumi-random==4.14.0
pulumi-cloudflare==5.13.0

deepmerge>=1.0.0,<2.0.0
docker[ssh]>=6.0.0,<7.0.0
httpx<1.0.0
tomlkit<1.0.0
jsonschema>=4.0.0,<5.0.0
pykeepass>=4.0.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
python-slugify>=8.0.0,<9.0.0
pycryptodome>=3.19.0,<4.0.0
filelock>=3.13.1,<4.0.0
argon2-cffi>=23.1.0,<24.0.0
